Testing Room Databases :)

+++ app/src/main/java/com/example/myunittestpracticeapp/data/local/ShoppingItem.kt
@@ -8,7 +8,7 @@ data class ShoppingItem(
 
     var name: String,
     var amount: Int,
-    var price: Long,
+    var price: Float,
     var imageUrl: String,
     @PrimaryKey(autoGenerate = true)
     val id: Int? = null
